covered old app installations as well
