+++ msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/testpass/broker/ltw/TestCase2579095.java
@@ -22,6 +22,8 @@
 //  THE SOFTWARE.
 package com.microsoft.identity.client.msal.automationapp.testpass.broker.ltw;
 
+import androidx.annotation.NonNull;
+
 import com.microsoft.identity.client.msal.automationapp.R;
 import com.microsoft.identity.client.msal.automationapp.testpass.broker.AbstractMsalBrokerTest;
 import com.microsoft.identity.client.ui.automation.annotations.LTWTests;
@@ -36,16 +38,37 @@ import com.microsoft.identity.client.ui.automation.interaction.PromptParameter;
 import com.microsoft.identity.client.ui.automation.interaction.microsoftsts.MicrosoftStsPromptHandlerParameters;
 import com.microsoft.identity.labapi.utilities.client.LabQuery;
 import com.microsoft.identity.labapi.utilities.constants.TempUserType;
+import com.microsoft.identity.labapi.utilities.constants.UserType;
 
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+import java.util.Arrays;
+import java.util.List;
 
 // Even Authenticator has the highest priority, if CP already has an artifact, CP will remain the broker.
 // https://identitydivision.visualstudio.com/Engineering/_workitems/edit/2579095
 @LTWTests
 @RunOnAPI29Minus
 @RetryOnFailure
+@RunWith(Parameterized.class)
 public class TestCase2579095 extends AbstractMsalBrokerTest {
+
+    private final UserType mUserType;
+
+    public TestCase2579095(@NonNull UserType userType) {
+        mUserType = userType;
+    }
+
+    @Parameterized.Parameters(name = "{0}")
+    public static List<UserType> userType() {
+        return Arrays.asList(
+                UserType.MSA,
+                UserType.CLOUD
+        );
+    }
 
     @Test
     public void test_2579095() throws Throwable {
@@ -65,7 +88,7 @@ public class TestCase2579095 extends AbstractMsalBrokerTest {
         final OneAuthTestApp oldOneAuthTestApp = new OneAuthTestApp();
         oldOneAuthTestApp.installOldApk();
         oldOneAuthTestApp.launch();
-        oldOneAuthTestApp.handleFirstRun();
+        oldOneAuthTestApp.handleFirstRunBasedOnUserType(mUserType);
 
         // acquire token interactively on OneAuthTestApp
         final FirstPartyAppPromptHandlerParameters promptHandlerParametersOneAuth = FirstPartyAppPromptHandlerParameters.builder()
@@ -92,7 +115,7 @@ public class TestCase2579095 extends AbstractMsalBrokerTest {
         // install new MsalTestApp
         msalTestApp.install();
         msalTestApp.launch();
-        msalTestApp.handleFirstRun();
+        msalTestApp.handleFirstRunBasedOnUserType(mUserType);
 
         // acquire token interactively on MsalTestApp and should not get prompt
         final MicrosoftStsPromptHandlerParameters promptHandlerParametersMsal = MicrosoftStsPromptHandlerParameters.builder()
@@ -126,12 +149,14 @@ public class TestCase2579095 extends AbstractMsalBrokerTest {
 
     @Override
     public LabQuery getLabQuery() {
-        return null;
+        return LabQuery.builder()
+                .userType(mUserType)
+                .build();
     }
 
     @Override
     public TempUserType getTempUserType() {
-        return TempUserType.BASIC;
+        return null;
     }
 
     @Override
