--- msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/testpass/msalonly/ltw/TestCase2517374.java
@@ -1,114 +1,0 @@
-//  Copyright (c) Microsoft Corporation.
-//  All rights reserved.
-//
-//  This code is licensed under the MIT License.
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files(the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions :
-//
-//  The above copyright notice and this permission notice shall be included in
-//  all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//  THE SOFTWARE.
-package com.microsoft.identity.client.msal.automationapp.testpass.msalonly.ltw;
-
-import com.microsoft.identity.client.msal.automationapp.AbstractMsalUiTest;
-import com.microsoft.identity.client.msal.automationapp.R;
-import com.microsoft.identity.client.msal.automationapp.testpass.broker.AbstractMsalBrokerTest;
-import com.microsoft.identity.client.ui.automation.annotations.LTWTests;
-import com.microsoft.identity.client.ui.automation.annotations.SupportedBrokers;
-import com.microsoft.identity.client.ui.automation.app.MsalTestApp;
-import com.microsoft.identity.client.ui.automation.broker.BrokerCompanyPortal;
-import com.microsoft.identity.client.ui.automation.interaction.PromptParameter;
-import com.microsoft.identity.client.ui.automation.interaction.microsoftsts.MicrosoftStsPromptHandlerParameters;
-import com.microsoft.identity.labapi.utilities.client.LabQuery;
-import com.microsoft.identity.labapi.utilities.constants.TempUserType;
-import org.junit.Assert;
-import org.junit.Test;
-import java.util.List;
-
-// Add a UI test on MsalTestApp
-// https://identitydivision.visualstudio.com/Engineering/_workitems/edit/2517374
-@LTWTests
-public class TestCase2517374 extends AbstractMsalUiTest {
-
-    @Test
-    public void test_2517374() throws Throwable {
-        final String username = mLabAccount.getUsername();
-        final String password = mLabAccount.getPassword();
-
-        MsalTestApp msalTestApp = new MsalTestApp();
-        msalTestApp.install();
-        msalTestApp.launch();
-        msalTestApp.handleFirstRun();
-
-        // acquire toke interactively and validate the token
-        final MicrosoftStsPromptHandlerParameters promptHandlerParameters = MicrosoftStsPromptHandlerParameters.builder()
-                .prompt(PromptParameter.SELECT_ACCOUNT)
-                .loginHint(username)
-                .sessionExpected(false)
-                .broker(null)
-                .expectingBrokerAccountChooserActivity(false)
-                .expectingProvidedAccountInBroker(false)
-                .expectingLoginPageAccountPicker(false)
-                .expectingProvidedAccountInCookie(false)
-                .consentPageExpected(false)
-                .passwordPageExpected(true)
-                .speedBumpExpected(false)
-                .registerPageExpected(false)
-                .enrollPageExpected(false)
-                .staySignedInPageExpected(false)
-                .verifyYourIdentityPageExpected(false)
-                .howWouldYouLikeToSignInExpected(false)
-                .build();
-
-        String token = msalTestApp.acquireToken(username, password, promptHandlerParameters, mBrowser, true, true);
-        Assert.assertNotNull(token);
-
-        // then acquire token silently and validate the token
-        msalTestApp.handleBackButton();
-        String silentToken = msalTestApp.acquireTokenSilent();
-        Assert.assertNotNull(silentToken);
-
-        // finally get users and validate the users
-        msalTestApp.handleBackButton();
-        List<String> users = msalTestApp.getUsers();
-        Assert.assertEquals(1, users.size());
-    }
-
-    // if getLabQuery return null then will use getTempUserType to create account
-    @Override
-    public LabQuery getLabQuery() {
-        return null;
-    }
-
-    @Override
-    public TempUserType getTempUserType() {
-        return TempUserType.BASIC;
-    }
-
-    @Override
-    public String[] getScopes() {
-        return new String[]{"User.read"};
-    }
-
-    @Override
-    public String getAuthority() {
-        return mApplication.getConfiguration().getDefaultAuthority().getAuthorityURL().toString();
-    }
-
-    @Override
-    public int getConfigFileResourceId() {
-        return R.raw.msal_config_default;
-    }
-}
